requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.19.4
Pillow>=8.0.1